requirements.txt
alembic==1.13.1
Authlib==1.3.0
autopep8==2.1.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
ecdsa==0.18.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
